Little float fixes & storage to update default value too

+++ Signals/Emitters/SignalFloat.cs
@@ -5,7 +5,7 @@ namespace Kalkatos.UnityGame.Scriptable
 	[CreateAssetMenu(fileName = "NewSignalFloat", menuName = "Signals/Signal (Float)", order = 6)]
 	public class SignalFloat : TypedSignal<float>
 	{
-		public int DefaultValue;
+		public float DefaultValue;
 
 		private float lastLogTime;
